cli/pyproject.toml
[project]
name = "llm-adventure-cli"
version = "0.1.0"
description = "LLM Adventure CLI - Interactive world-building interface"
requires-python = ">=3.11"
dependencies = [
    "httpx==0.27.0",
    "rich==13.7.0",
    "python-dotenv==1.0.0",
    "pydantic==2.9.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
adventure = "adventure_cli:main_menu"
